3lab/requirements.txt
#requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0.3
celery>=5.3.0
redis>=4.5.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
eventlet>=0.33.0,<0.41.0  # Исправленная версия
python-multipart
sqlalchemy>=2.0.0
alembic>=1.15.2
bcrypt>=4.3.0
cryptography>=44.0.2
python-jose>=3.4.0
passlib>=1.7.4
annotated-types>=0.7.0
anyio>=4.9.0
click>=8.1.8
colorama>=0.4.6
ecdsa>=0.19.1
exceptiongroup>=1.2.2
greenlet>=3.1.1
h11>=0.14.0
idna>=3.10
mako>=1.3.9
markupsafe>=3.0.2
pyasn1>=0.4.8
pycparser>=2.22
pydantic-core>=2.33.1
rsa>=4.9
six>=1.17.0
sniffio>=1.3.1
starlette>=0.46.1
typing-inspection>=0.4.0
typing-extensions>=4.13.1
httpx>=0.27.0